refactored FKBP displacement workflow and moved helper function to AtomUtils
